requirements.txt
SQLAlchemy
PyMySQL
pandas
python-dotenv
